examples/requirements.txt
streamlit==1.31.0
langchain-community==0.2.16 
langchain-text-splitters==0.2.4 
langchain-chroma==0.1.3 
sentence_transformers==3.0.1 
langchain_openai==0.1.23
beautifulsoup4==4.12.3